+++ packages/media-parser/src/state/parser-state.ts
@@ -8,17 +8,11 @@ import {
 } from '../boxes/webm/traversal';
 import type {BufferIterator} from '../buffer-iterator';
 import type {Options, ParseMediaFields} from '../options';
-import type {
-	AudioOrVideoSample,
-	OnAudioSample,
-	OnAudioTrack,
-	OnVideoSample,
-	OnVideoTrack,
-} from '../webcodec-sample-types';
+import type {OnAudioTrack, OnVideoTrack} from '../webcodec-sample-types';
 import {makeCanSkipTracksState} from './can-skip-tracks';
-import {makeTracksSectionState} from './has-tracks-section';
 import {keyframesState} from './keyframes';
 import {riffSpecificState} from './riff';
+import {sampleCallback} from './sample-callbacks';
 import {structureState} from './structure';
 
 export type InternalStats = {
@@ -82,12 +76,6 @@ export const makeParserState = ({
 		};
 	};
 
-	const videoSampleCallbacks: Record<number, OnVideoSample> = {};
-	const audioSampleCallbacks: Record<number, OnAudioSample> = {};
-
-	const queuedAudioSamples: Record<number, AudioOrVideoSample[]> = {};
-	const queuedVideoSamples: Record<number, AudioOrVideoSample[]> = {};
-
 	let timescale: number | null = null;
 	let skippedBytes: number = 0;
 
@@ -133,111 +121,29 @@ export const makeParserState = ({
 		return timestampMap.get(byteOffset);
 	};
 
-	const samplesForTrack: Record<number, number> = {};
-
-	const avcProfile: SpsAndPps | null = null;
-
 	const canSkipTracksState = makeCanSkipTracksState({
 		hasAudioTrackHandlers,
 		fields,
 		hasVideoTrackHandlers,
 	});
-	const tracksState = makeTracksSectionState(canSkipTracksState);
 	const riff = riffSpecificState();
+	const sample = sampleCallback(canSkipTracksState, signal);
 
 	return {
 		riff,
+		sample,
 		onTrackEntrySegment,
 		getTrackInfoByNumber: (id: number) => trackEntries[id],
-		registerVideoSampleCallback: async (
-			id: number,
-			callback: OnVideoSample | null,
-		) => {
-			if (callback === null) {
-				delete videoSampleCallbacks[id];
-				return;
-			}
-
-			videoSampleCallbacks[id] = callback;
-
-			for (const queued of queuedVideoSamples[id] ?? []) {
-				await callback(queued);
-			}
-
-			queuedVideoSamples[id] = [];
-		},
 		setTimestampOffset,
 		getTimestampOffsetForByteOffset,
-		registerAudioSampleCallback: async (
-			id: number,
-			callback: OnAudioSample | null,
-		) => {
-			if (callback === null) {
-				delete audioSampleCallbacks[id];
-				return;
-			}
-
-			audioSampleCallbacks[id] = callback;
-			for (const queued of queuedAudioSamples[id] ?? []) {
-				await callback(queued);
-			}
-
-			queuedAudioSamples[id] = [];
-		},
-		onAudioSample: async (trackId: number, audioSample: AudioOrVideoSample) => {
-			if (signal?.aborted) {
-				throw new Error('Aborted');
-			}
-
-			if (typeof samplesForTrack[trackId] === 'undefined') {
-				samplesForTrack[trackId] = 0;
-			}
-
-			samplesForTrack[trackId]++;
-
-			const callback = audioSampleCallbacks[trackId];
-			if (callback) {
-				await callback(audioSample);
-			}
-		},
-		onVideoSample: async (trackId: number, videoSample: AudioOrVideoSample) => {
-			if (signal?.aborted) {
-				throw new Error('Aborted');
-			}
-
-			if (typeof samplesForTrack[trackId] === 'undefined') {
-				samplesForTrack[trackId] = 0;
-			}
-
-			samplesForTrack[trackId]++;
-
-			const callback = videoSampleCallbacks[trackId];
-			if (callback) {
-				await callback(videoSample);
-			}
-		},
 		getTimescale,
 		setTimescale,
-		getSamplesForTrack: (trackId: number) => {
-			return samplesForTrack[trackId] ?? 0;
-		},
-		getAvcProfile: () => {
-			return avcProfile;
-		},
 		getInternalStats: (): InternalStats => ({
 			skippedBytes,
 			finalCursorOffset: getIterator()?.counter.getOffset() ?? 0,
 		}),
 		getSkipBytes: () => skippedBytes,
 		increaseSkippedBytes,
-		maySkipVideoData: () => {
-			return (
-				tracksState.hasAllTracks() &&
-				Object.values(videoSampleCallbacks).length === 0 &&
-				Object.values(audioSampleCallbacks).length === 0
-			);
-		},
-		tracks: tracksState,
 		canSkipTracksState,
 		keyframes,
 		structure,
